qmanual/pyproject.toml
[project]
name = "qmanual"
version = "0.1.0"
description = "Generate Quarto manual scaffolds (website/book) from YAML spec."
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "jinja2>=3.1",
]

[project.scripts]
qmanual = "qmanual.cli:main"

[tool.setuptools.package-data]
qmanual = ["templates/*.j2"]
